backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.6,<1.0.0

# Database
asyncpg==0.29.0
pgvector==0.2.4

# HTTP Client
httpx==0.27.0

# Configuration & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
supabase==2.3.0
gotrue==2.1.0
python-jose[cryptography]==3.3.0

# Scientific Computing
numpy==1.26.3
umap-learn==0.5.5
hdbscan==0.8.33
scikit-learn==1.4.0
scipy==1.12.0

# Caching
redis==5.0.1

# Retry Logic
tenacity==8.2.3
